requirements.txt
openai==2.9.0
python-dotenv==1.2.1
httpx==0.28.1
faiss-cpu==1.13.0
sentence-transformers==2.6.1
transformers==4.37.2
torch==2.2.1
